Split widget: effects header
